requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
bs4==0.0.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
commentjson==0.9.0
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.13
docker-pycreds==0.4.0
einops==0.6.1
fastjsonschema==2.19.0
filelock==3.12.0
Flask==3.0.3
fonttools==4.39.4
fqdn==1.5.1
freetype-py==2.4.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-resources==5.12.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
justpfm==1.0.0
kiwisolver==1.4.4
kornia==0.6.12
lark-parser==0.7.8
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.3
lpips==0.1.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
networkx==3.1
notebook==7.0.7
notebook_shim==0.2.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.7.0.72
OpenEXR==3.2.4
overrides==7.7.0
pandas==2.0.1
pandocfilters==1.5.0
pathtools==0.1.2
pickle5==0.0.11
Pillow==9.5.0
plotly==5.22.0
plyfile==0.9
prometheus-client==0.19.0
protobuf==3.20.1
pvlib==0.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycuda==2022.2.2
pyDeprecate==0.3.2
pyglet==1.5.27
pymesh==1.0.2
pymeshlab==0.2.1
PyOpenGL==3.1.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrender==0.1.45
python-json-logger==2.0.7
pytools==2022.1.14
pytorch-lightning==1.8.5
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
rsa==4.9
Rtree==1.0.1
scikit-image==0.20.0
scikit-learn==1.4.2
scipy==1.9.1
Send2Trash==1.8.2
sentry-sdk==1.22.2
setproctitle==1.3.2
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.12
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.5.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
trimesh==3.21.6
triton==2.0.0
types-python-dateutil==2.8.19.20240106
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
utm==0.7.0
vedo==2023.4.4
vtk==9.2.6
wandb==0.16.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.15.0
yarl==1.9.2
zipp==3.13.0